requirements.txt
python-telegram-bot==12.8
requests==2.24.0
selectorlib==0.16.0

pytest==6.0.2
prospector==1.3.0
